Project fixes under MacOS
Former-commit-id: a04e11fbd321c8f8fe3432073c6b6c62f0a39322

+++ Sources/Internal/Base/Introspection.h
@@ -95,6 +95,6 @@ namespace DAVA
 	}
 
 #define MEMBER(_name, _desc) \
-	DAVA::IntrospectionMember(#_name, _desc, (int) &((ObjectT *) 0)->_name, DAVA::MetaInfo::Instance(&ObjectT::_name)),
+	DAVA::IntrospectionMember(#_name, _desc, (int) ((long int) &((ObjectT *) 0)->_name), DAVA::MetaInfo::Instance(&ObjectT::_name)),
 
 #endif // __DAVAENGINE_INTROSPECTION_H__
